# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/storage/db/FilmDbStorage.java

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.interfaces.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.interfaces.UserStorage;
 
@@ -13,20 +14,27 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Slf4j
 @Service
 public class FilmService {
     private final UserStorage userStorage;
     private final FilmStorage filmStorage;
-    private static final LocalDate LIMIT_DATE = LocalDate.from(LocalDateTime.of(1895, 12, 28, 0, 0));
+    private static final LocalDate LIMIT_DATE = LocalDate.from(
+            LocalDateTime.of(1895, 12, 28, 0, 0));
     private static final int LIMIT_LENGTH_OF_DESCRIPTION = 200;
 
     @Autowired
     public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
         this.userStorage = userStorage;
         this.filmStorage = filmStorage;
+    }
+
+    public List<Film> getDirectorsFilmsSortedBy(int directorId, String sortBy) {
+        return filmStorage.getDirectorsFilms(directorId, sortBy);
     }
 
     public Collection<Film> getAll() {
@@ -48,12 +56,19 @@ public class FilmService {
         return result;
     }
 
-    public void delete(int filmId) {
+    public Film delete(Integer filmId) {
+        log.info("Request to delete the movie by ID = " + filmId + " received");
+        if (filmId == null) {
+            throw new NotFoundException("Movie with ID = " + filmId + " not found");
+        }
+        if (filmId < 0) {
+            throw new NotFoundException("Movie with ID = " + filmId + " not found");
+        }
         if (getById(filmId) == null) {
             throw new NotFoundException("Movie with ID = " + filmId + " not found");
         }
         log.info("Deleted film with id: {}", filmId);
-        filmStorage.delete(filmId);
+        return filmStorage.delete(filmId);
     }
 
     public Film getById(Integer id) {
@@ -63,16 +78,22 @@ public class FilmService {
 
     public void addLike(Integer filmId, Integer userId) {
         Film film = filmStorage.getById(filmId);
-        if (film != null) {
-            if (userStorage.getById(userId) != null) {
-                filmStorage.addLike(filmId, userId);
-                log.info("Like successfully added");
-            } else {
-                throw new NotFoundException("User with ID = " + userId + " not found");
-            }
-        } else {
+        if (film == null) {
             throw new NotFoundException("Movie with ID = " + filmId + " not found");
         }
+        if (userStorage.getById(userId) == null) {
+            throw new NotFoundException("User with ID = " + userId + " not found");
+        }
+        filmStorage.addLike(filmId, userId);
+        User user = userStorage.getById(userId);
+        Set<Integer> userLikes = user.getLikes();
+        if (userLikes == null) {
+            userLikes = new HashSet<>(filmId);
+        }
+        userLikes.add(filmId);
+        user.setLikes(userLikes);
+        userStorage.update(user);
+        log.info("Like successfully added");
     }
 
     public void removeLike(Integer filmId, Integer userId) {
@@ -95,8 +116,9 @@ public class FilmService {
         return result;
     }
 
-    protected void validate(Film film, String message) {
-        if (film.getDescription().length() > LIMIT_LENGTH_OF_DESCRIPTION || film.getReleaseDate().isBefore(LIMIT_DATE)) {
+    private void validate(Film film, String message) {
+        if (film.getDescription().length() > LIMIT_LENGTH_OF_DESCRIPTION || film.getReleaseDate()
+                .isBefore(LIMIT_DATE)) {
             log.debug(message);
             throw new ValidationException(message);
         }
